+++ core/middleware.py
@@ -1,6 +1,8 @@
 import sigauth.middleware
 
 from django.conf import settings
+from django.utils.deprecation import MiddlewareMixin
+from django.http import HttpResponse
 
 
 class SignatureCheckMiddleware(
@@ -14,3 +16,11 @@ class SignatureCheckMiddleware(
         ] or request.path_info.startswith('/admin/login'):
             return False
         return super().should_check(request)
+
+
+class CheckStaffStatusMiddleware(MiddlewareMixin):
+
+    def process_view(self, request, view_func, view_args, view_kwarg):
+        if request.user.is_authenticated():
+            if not request.user.is_staff:
+                return HttpResponse('User not authorized for admin access', status=401)